# backend/src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Register API Documentation",
    "description": "This is the documentation for the Register API",
    "termsOfService": "http://swagger.io/terms/",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://localhost:5000",
      "description": "API server"
    }
  ],
  "paths": {
    "/register/": {
      "post": {
        "summary": "Register a new user",
        "description": "Register a new user",
        "tags": [
          "register"
        ],
        "operationId": "register",
        "requestBody": {
          "description": "User object that needs to be added to the store",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              },
              "examples": {
                "user": {
                  "value": {
                    "name": "Jon Doe",
                    "email": "jondoe@email.com",
                    "telephone": "9999-9999",
                    "address": "Avenida Jon Doe",
                    "cpf": "Jon"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/register":{
      "get": {
        "summary": "Get all users",
        "description": "Get all users",
        "tags": [
          "register"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "operationId": "getUsers",
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                },
                "examples": {
                  "user": {
                    "value": {
                      "name": "Jon Doe",
                      "email": "jondoe@email.com",
                      "telephone": "9999-9999",
                      "address": "Avenida Jon Doe",
                      "cpf": "Jon"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/register/{id}/":{
      "put": {
        "summary": "Update a user",
        "description": "Update a user",
        "tags": [
          "register"
        ],
        "operationId": "updateUser",
        "requestBody": {
          "description": "User object that needs to be added to the store",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              },
              "examples": {
                "user": {
                  "value": {
                    "name": "Jon Doe",
                    "email": "jondoe@email.com",
                    "telephone": "9999-9999",
                    "address": "Avenida Jon Doe",
                    "cpf": "Jon"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/register/{id}": {
      "delete": {
        "summary": "Delete a user",
        "description": "Delete a user",
        "tags": [
          "register"
        ],
        "operationId": "deleteUser",
        "requestBody": {
          "description": "User object that needs to be added to the store",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              },
              "examples": {
                "user": {
                  "value": {
                    "name": "Jon Doe",
                    "email": "jondoe@email.com",
                    "telephone": "9999-9999",
                    "address": "Avenida Jon Doe",
                    "cpf": "Jon"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/login/": {
      "post": {
        "summary": "Login a user",
        "description": "Login a user",
        "tags": [
          "login"
        ],
        "operationId": "login",
        "requestBody": {
          "description": "User object that needs to be added to the store",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              },
              "examples": {
                "user": {
                  "value": {
                    "username": "JonDoe",
                    "password": "********"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/login":{
      "get": {
        "summary": "Get all users",
        "description": "Get all users",
        "tags": [
          "login"
        ],
        "operationId": "getUsers",
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                },
                "examples": {
                  "user": {
                    "value": {
                      "username": "JonDoe",
                      "password": "********"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/login/{id}":{
      "delete": {
        "summary": "Delete a user",
        "description": "Delete a user",
        "tags": [
          "login"
        ],
        "operationId": "deleteUser",
        "requestBody": {
          "description": "User object that needs to be added to the store",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              },
              "examples": {
                "user": {
                  "value": {
                    "username": "JonDoe",
                    "password": "********"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/auth": {
      "post": {
        "summary": "Authenticate a user",
        "description": "Authenticate a user",
        "tags": [
          "auth"
        ],
        "operationId": "auth",
        "requestBody": {
          "description": "User object that needs to be added to the store",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "username": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "userStatus": {
            "type": "integer",
            "format": "int32",
            "description": "User Status"
          }
        }
      }
    }
  }
}